Adding normal DHCP and ZTP
adding openconfig backend

managing the template and external files

interface to template mapping for ztp

switching to use the inventory client interface

using template from citadel for dhcp

In order to detact this pkg library from Citadel, a new Inventory interface is
use to abstract the client access to the Inventory backend. However, lots of
templating related data and structures are still being used here.

dhcp to get ip from topo

supporting ip addressing from openconfig

correctly handle the url request

right now, cumulus ztp (possibly a bug?) when passed with query parameters, it
insert escape characters by default but did not consumed it so it ended up
sending the url with the embeded escape characters.

since it's invalid from the perspective of the http server as a request url, we
need to strip it out